requirements.txt
Django==1.7.1
Pillow==2.5.3
djangorestframework==3.0.2
wsgiref==0.1.2
